cleaned hand gesture branch and added DecisionFilter
